requirements.txt
# GEC Rajkot Backend Dependencies
# Flask Framework and Extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.21

# Security and Authentication
Werkzeug==2.3.7
PyJWT==2.8.0
bcrypt==4.0.1

# Email and OTP
email-validator==2.0.0

# Google API for Email Notifications
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0

# Development Dependencies (optional)
pytest==7.4.2
pytest-flask==1.2.0
coverage==7.3.2

# Production Dependencies (optional)
gunicorn==21.2.0
psycopg2-binary==2.9.7  # For PostgreSQL in production
sendgrid==6.11.0
Pillow==10.1.0

# Excel file processing
openpyxl==3.1.2
pandas==2.1.1